pyproject.toml
[tool.poetry]
name = "eufy-clean"
version = "0.1.0"
description = ""
authors = ["Tobias Jeppesen <github@tobiasjeppesen.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pycryptodomex = "^3.21.0"
protobuf = "^5.29.3"
paho-mqtt = "^2.1.0"
dnspython = "^2.7.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
